move openapi module to core
